remove support for generic listeners for Push events
