+++ src/main/java/redis/clients/jedis/ConnectionFactory.java
@@ -26,7 +26,6 @@ public class ConnectionFactory implements PooledObjectFactory<Connection> , Rebi
     private Connection.Builder connectionBuilder;
     private JedisSocketFactory jedisSocketFactory;
     private Cache cache;
-    private PushHandler pushHandler;
     private HostAndPort hostAndPort;
 
     // Fluent API methods (preferred)
@@ -48,11 +47,6 @@ public class ConnectionFactory implements PooledObjectFactory<Connection> , Rebi
     public Builder cache(Cache cache) {
       this.cache = cache;
       return this;
-    }
-
-    public Builder pushHandler(PushHandler pushHandler) {
-        this.pushHandler = pushHandler;
-        return this;
     }
 
     public Builder hostAndPort(HostAndPort hostAndPort) {
