--- src/main/java/redis/clients/jedis/ListenerHandler.java
@@ -1,14 +1,0 @@
-package redis.clients.jedis;
-
-import java.util.Collection;
-
-public interface ListenerHandler<T> {
-  void addListener(T listener);
-
-  void removeListener(T listener);
-
-  void removeAllListeners();
-
-  Collection<T> getListeners();
-
-}
