+++ src/main/java/redis/clients/jedis/MaintenanceEventHandler.java
@@ -1,5 +1,26 @@
 package redis.clients.jedis;
 
-public interface MaintenanceEventHandler extends ListenerHandler<MaintenanceEventListener> {
+import java.util.Collection;
+import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
 
+public class MaintenanceEventHandler {
+
+  private final List<MaintenanceEventListener> listeners = new CopyOnWriteArrayList<>();
+
+  public void addListener(MaintenanceEventListener listener) {
+    listeners.add(listener);
+  }
+
+  public void removeListener(MaintenanceEventListener listener) {
+    listeners.remove(listener);
+  }
+
+  public void removeAllListeners() {
+    listeners.clear();
+  }
+
+  public Collection<MaintenanceEventListener> getListeners() {
+    return listeners;
+  }
 }
