--- src/main/java/redis/clients/jedis/MaintenanceEventHandlerImpl.java
@@ -1,5 +1,0 @@
-package redis.clients.jedis;
-
-public class MaintenanceEventHandlerImpl extends AbstractListenerHandler<MaintenanceEventListener>
-    implements MaintenanceEventHandler {
-}
