--- src/main/java/redis/clients/jedis/PushHandelrImpl.java
@@ -1,4 +1,0 @@
-package redis.clients.jedis;
-
-class PushHandlerImpl extends AbstractListenerHandler<PushListener> implements PushHandler {
-}
